+++ docs/snippets/coordination/kubernetes.py
@@ -1,7 +1,7 @@
 from grelmicro import Grelmicro
 from grelmicro.coordination import Coordination
 from grelmicro.coordination.kubernetes import (
-    KubernetesLeaderElectionBackend,
+    KubernetesLeaderElectionAdapter,
     KubernetesLockAdapter,
 )
 
@@ -9,7 +9,7 @@ micro = Grelmicro(
     uses=[
         Coordination(
             lock=KubernetesLockAdapter(namespace="default"),
-            election=KubernetesLeaderElectionBackend(namespace="default"),
+            election=KubernetesLeaderElectionAdapter(namespace="default"),
         )
     ]
 )