+++ tests/task/conftest.py
@@ -8,7 +8,7 @@ import pytest
 
 from grelmicro.coordination.abc import LeaderElectionBackend, LockBackend
 from grelmicro.coordination.memory import (
-    MemoryLeaderElectionBackend,
+    MemoryLeaderElectionAdapter,
     MemoryLockAdapter,
 )
 from grelmicro.coordination.tasklock import TaskLock
@@ -87,8 +87,8 @@ async def backend() -> AsyncGenerator[LockBackend]:
 
 @pytest.fixture
 async def leader_backend() -> AsyncGenerator[LeaderElectionBackend]:
-    """Return Memory Leader Election Backend."""
-    async with MemoryLeaderElectionBackend() as backend:
+    """Return Memory Leader Election Adapter."""
+    async with MemoryLeaderElectionAdapter() as backend:
         yield backend